+++ NEWS.md
@@ -1,3 +1,9 @@
+OhidsShinyAppBuilder v1.1.0
+=============================
+- Moved the protocol link to a button on the sidebar 
+- Moved the study description from a top div in all modules to a button with a pop up modal in the sidebar
+- Added option to add html into app header (e.g., for tracking user access)
+
 OhidsShinyAppBuilder v1.0.0
 =============================
 Changed name from ShinyAppBuilder and created a new repo with the new name that copies https://github.com/OHDSI/ShinyAppBuilder